+++ includes/wp-admin/class-blog-settings-fields.php
@@ -5,7 +5,7 @@
  * @package ActivityPub
  */
 
-namespace Activitypub;
+namespace Activitypub\WP_Admin;
 
 /**
  * Class to handle all blog settings fields and callbacks.
@@ -15,7 +15,7 @@ class Blog_Settings_Fields {
 	 * Initialize the settings fields.
 	 */
 	public static function init() {
-		add_action( 'settings_page_activitypub', array( self::class, 'register_settings' ) );
+		add_action( 'load-settings_page_activitypub', array( self::class, 'register_settings' ) );
 	}
 
 	/**
@@ -143,7 +143,7 @@ class Blog_Settings_Fields {
 	public static function profile_id_callback() {
 		?>
 		<label for="activitypub_blog_identifier">
-			<input id="activitypub_blog_identifier" class="blog-user-identifier" name="activitypub_blog_identifier" type="text" value="<?php echo esc_attr( get_option( 'activitypub_blog_identifier', Model\Blog::get_default_username() ) ); ?>" />
+			<input id="activitypub_blog_identifier" class="blog-user-identifier" name="activitypub_blog_identifier" type="text" value="<?php echo esc_attr( get_option( 'activitypub_blog_identifier', \Activitypub\Model\Blog::get_default_username() ) ); ?>" />
 			@<?php echo esc_html( wp_parse_url( home_url(), PHP_URL_HOST ) ); ?>
 		</label>
 		<p class="description">
@@ -188,7 +188,7 @@ class Blog_Settings_Fields {
 
 		<table class="widefat striped activitypub-extra-fields" role="presentation" style="margin: 15px 0;">
 		<?php
-		$extra_fields = Collection\Extra_Fields::get_actor_fields( Collection\Actors::BLOG_USER_ID );
+		$extra_fields = \Activitypub\Collection\Extra_Fields::get_actor_fields( \Activitypub\Collection\Actors::BLOG_USER_ID );
 
 		if ( empty( $extra_fields ) ) :
 			?>
